archive/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
anthropic==0.43.0
pydantic==2.10.4
pytest==8.3.4
httpx==0.28.1
mcp>=1.22.0
